Fixed env_var parsing
It is a little different, as it is defined positively instead of negatively like the rest of the tokens.

Anything else than alphanum or `_' will be seen as the end of the environment variable.
If the environment variable does not start with an aplhanum or `_' character, the token is treated as a regular string